requirements.txt
opencv-python
matplotlib
torch
torchvision
Pillow
tensorboard
streamlit
tqdm
scikit-learn